remove theme toggle initial flash when dark theme
ssr renders the toggle as unchecked, but we run some specific code in the head to ensure that the <html> element has its theme attr set early. it takes one render though for the checkbox to become checked if the theme should be dark, so until we've loaded that state in and re-rendered the toggle, we display it as checked (the dark theme case) if the html theme attr is set.