Split Streams API documentation into multiple pages.
The Streams API documentation was originally deeply nested, like so:

* Streams:
** Background
** Types
*** TransformStream
**** Constructor
**** Methods
... etc

This made it hard to follow, and also hard to mark up and style effectively.

This commit separates each of the "Types" into its own Streams API page.
This makes sense because each type has a constructor, methods, properties,
etc.

In doing so, it also made sense to pull the introduction to Streams, and
why and how to use them, into the Learning section rather than the
Reference section.

Additionally, I've added more example usage and links.