Update ESLint config and make it run during development
Generally follows the docs for setting up custom ESLint rules with
Gatsby: https://www.gatsbyjs.org/docs/eslint/

Specifically, when we include our own .eslintrc or .eslintrc.js file,
it overrides the built-in ESLint configuration that comes with Gatsby.

In order to re-enable running ESLint automatically on build during
development, we need to either use `gatsby-plugin-eslint` or set up
`eslint-loader` and the default rules manually. I chose to use the
plugin.

This change surfaces 1 ESLint error and a handful of warnings, which
will be fixed in the following commit.

* Adds an .eslintrc.js that follows