Simplify path prefix handling
`gatsby develop`, the command for building and serving the Gatsby app locally for development, does not support the `--prefix-paths` flag, meaning that it can't re-route everything behind the `assetPrefix` value (in our site, `/workers`). Regardless of that, the links _do_ still get the `/workers` value prefixed to them, but the project still builds out to a base path of `/`.

To handle this, we can check for the NODE_ENV environment variable, and set the `assetPrefix` based on that value. If we're building the project for production, we can set the `assetPrefix` to `/workers`, allowing it to be served correctly from that subpath. If we're in development, we shouldn't have an `assetPrefix`, allowing the application to be rendered at `/` and have correct active sidebar link semantics, and everything else around the routing.