fix sizing of small tables inside table wraps
`min-width: 100%` was originally working in chrome at the time <TableWrap/> was introduced. it still works (as of 2020-10-29) in safari and partially works in ff. table sizing behavior is notoriously tricky, so we use calc and absolute units (subtracting from 100vw) to ensure the min-width is adhered to. this is a "nice-to-have" design thing so no fallback is necessary.