Pin color-string to fix problematic new release (#373)
The color-string package introduced a breaking change in a patch update in version [1.7.1](https://github.com/Qix-/color-string/releases/tag/1.7.1). The code in question changes the regular expression for keyword detection to a more permissive word-character match from a number character match.

The solution is to attempt to pin color-string to a previous working version.